+++ test/unit/v1/index.js
@@ -12,10 +12,12 @@ import terminalDirectory from './terminal_directory';
 import terminalV1 from './terminal_v1';
 import ticketBooth from './ticket_booth';
 
+import { snapshot, restore } from "../../utils"
+
 let snapshotId;
 export default function () {
   beforeEach(async function () {
-    snapshotId = await this.snapshotFn();
+    snapshotId = await snapshot();
     // Mark the start time of each test.
     await this.setTimeMarkFn();
   });
@@ -42,6 +44,6 @@ export default function () {
 
   // After each test, restore the contract state.
   afterEach(async function () {
-    await this.restoreFn(snapshotId);
+    await restore(snapshotId);
   });
 }
